feat: conslidate CategoryLinks component
